refactor adapters to fit more common use cases
